@@ -1,26 +1,26 @@
 {
-  "_from": "body-parser@1.19.0",
-  "_id": "body-parser@1.19.0",
+  "_from": "body-parser@1.19.1",
+  "_id": "body-parser@1.19.1",
   "_inBundle": false,
-  "_integrity": "sha512-dhEPs72UPbDnAQJ9ZKMNTP6ptJaionhP5cBb541nXPlW60Jepo9RV/a4fX4XWW9CuFNK22krhrj1+rgzifNCsw==",
+  "_integrity": "sha512-8ljfQi5eBk8EJfECMrgqNGWPEY5jWP+1IzkzkGdFFEwFQZZyaZ21UqdaHktgiMlH0xLHqIFtE/u2OYE5dOtViA==",
   "_location": "/body-parser",
   "_phantomChildren": {},
   "_requested": {
     "type": "version",
     "registry": true,
-    "raw": "body-parser@1.19.0",
+    "raw": "body-parser@1.19.1",
     "name": "body-parser",
     "escapedName": "body-parser",
-    "rawSpec": "1.19.0",
+    "rawSpec": "1.19.1",
     "saveSpec": null,
-    "fetchSpec": "1.19.0"
+    "fetchSpec": "1.19.1"
   },
   "_requiredBy": [
     "/express"
   ],
-  "_resolved": "https://registry.npmjs.org/body-parser/-/body-parser-1.19.0.tgz",
-  "_shasum": "96b2709e57c9c4e09a6fd66a8fd979844f69f08a",
-  "_spec": "body-parser@1.19.0",
+  "_resolved": "https://registry.npmjs.org/body-parser/-/body-parser-1.19.1.tgz",
+  "_shasum": "1499abbaa9274af3ecc9f6f10396c995943e31d4",
+  "_spec": "body-parser@1.19.1",
   "_where": "C:\\Users\\ronid\\Documents\\CS Final Project\\frontend\\node_modules\\express",
   "bugs": {
     "url": "https://github.com/expressjs/body-parser/issues"
@@ -38,32 +38,32 @@
     }
   ],
   "dependencies": {
-    "bytes": "3.1.0",
+    "bytes": "3.1.1",
     "content-type": "~1.0.4",
     "debug": "2.6.9",
     "depd": "~1.1.2",
-    "http-errors": "1.7.2",
+    "http-errors": "1.8.1",
     "iconv-lite": "0.4.24",
     "on-finished": "~2.3.0",
-    "qs": "6.7.0",
-    "raw-body": "2.4.0",
-    "type-is": "~1.6.17"
+    "qs": "6.9.6",
+    "raw-body": "2.4.2",
+    "type-is": "~1.6.18"
   },
   "deprecated": false,
   "description": "Node.js body parsing middleware",
   "devDependencies": {
-    "eslint": "5.16.0",
-    "eslint-config-standard": "12.0.0",
-    "eslint-plugin-import": "2.17.2",
-    "eslint-plugin-markdown": "1.0.0",
-    "eslint-plugin-node": "8.0.1",
-    "eslint-plugin-promise": "4.1.1",
-    "eslint-plugin-standard": "4.0.0",
-    "istanbul": "0.4.5",
+    "eslint": "7.32.0",
+    "eslint-config-standard": "14.1.1",
+    "eslint-plugin-import": "2.25.3",
+    "eslint-plugin-markdown": "2.2.1",
+    "eslint-plugin-node": "11.1.0",
+    "eslint-plugin-promise": "5.2.0",
+    "eslint-plugin-standard": "4.1.0",
     "methods": "1.1.2",
-    "mocha": "6.1.4",
-    "safe-buffer": "5.1.2",
-    "supertest": "4.0.2"
+    "mocha": "9.1.3",
+    "nyc": "15.1.0",
+    "safe-buffer": "5.2.1",
+    "supertest": "6.1.6"
   },
   "engines": {
     "node": ">= 0.8"
@@ -82,10 +82,10 @@
     "url": "git+https://github.com/expressjs/body-parser.git"
   },
   "scripts": {
-    "lint": "eslint --plugin markdown --ext js,md .",
+    "lint": "eslint .",
     "test": "mocha --require test/support/env --reporter spec --check-leaks --bail test/",
-    "test-cov": "istanbul cover node_modules/mocha/bin/_mocha -- --require test/support/env --reporter dot --check-leaks test/",
-    "test-travis": "istanbul cover node_modules/mocha/bin/_mocha --report lcovonly -- --require test/support/env --reporter spec --check-leaks test/"
+    "test-ci": "nyc --reporter=lcov --reporter=text npm test",
+    "test-cov": "nyc --reporter=html --reporter=text npm test"
   },
-  "version": "1.19.0"
+  "version": "1.19.1"
 }
