@@ -1,26 +1,26 @@
 {
-  "_from": "cookie@0.4.0",
-  "_id": "cookie@0.4.0",
+  "_from": "cookie@0.4.1",
+  "_id": "cookie@0.4.1",
   "_inBundle": false,
-  "_integrity": "sha512-+Hp8fLp57wnUSt0tY0tHEXh4voZRDnoIrZPqlo3DPiI4y9lwg/jqx+1Om94/W6ZaPDOUbnjOt/99w66zk+l1Xg==",
+  "_integrity": "sha512-ZwrFkGJxUR3EIoXtO+yVE69Eb7KlixbaeAWfBQB9vVsNn/o+Yw69gBWSSDK825hQNdN+wF8zELf3dFNl/kxkUA==",
   "_location": "/cookie",
   "_phantomChildren": {},
   "_requested": {
     "type": "version",
     "registry": true,
-    "raw": "cookie@0.4.0",
+    "raw": "cookie@0.4.1",
     "name": "cookie",
     "escapedName": "cookie",
-    "rawSpec": "0.4.0",
+    "rawSpec": "0.4.1",
     "saveSpec": null,
-    "fetchSpec": "0.4.0"
+    "fetchSpec": "0.4.1"
   },
   "_requiredBy": [
     "/express"
   ],
-  "_resolved": "https://registry.npmjs.org/cookie/-/cookie-0.4.0.tgz",
-  "_shasum": "beb437e7022b3b6d49019d088665303ebe9c14ba",
-  "_spec": "cookie@0.4.0",
+  "_resolved": "https://registry.npmjs.org/cookie/-/cookie-0.4.1.tgz",
+  "_shasum": "afd713fe26ebd21ba95ceb61f9a8116e50a537d1",
+  "_spec": "cookie@0.4.1",
   "_where": "C:\\Users\\ronid\\Documents\\CS Final Project\\frontend\\node_modules\\express",
   "author": {
     "name": "Roman Shtylman",
@@ -41,10 +41,10 @@
   "devDependencies": {
     "beautify-benchmark": "0.2.4",
     "benchmark": "2.1.4",
-    "eslint": "5.16.0",
-    "eslint-plugin-markdown": "1.0.0",
-    "istanbul": "0.4.5",
-    "mocha": "6.1.4"
+    "eslint": "6.8.0",
+    "eslint-plugin-markdown": "1.0.2",
+    "mocha": "7.1.1",
+    "nyc": "15.0.1"
   },
   "engines": {
     "node": ">= 0.6"
@@ -69,10 +69,10 @@
   "scripts": {
     "bench": "node benchmark/index.js",
     "lint": "eslint --plugin markdown --ext js,md .",
-    "test": "mocha --reporter spec --bail --check-leaks test/",
-    "test-ci": "istanbul cover node_modules/mocha/bin/_mocha --report lcovonly -- --reporter spec --check-leaks test/",
-    "test-cov": "istanbul cover node_modules/mocha/bin/_mocha -- --reporter dot --check-leaks test/",
+    "test": "mocha --reporter spec --bail --check-leaks --ui qunit test/",
+    "test-ci": "nyc --reporter=text npm test",
+    "test-cov": "nyc --reporter=html --reporter=text npm test",
     "version": "node scripts/version-history.js && git add HISTORY.md"
   },
-  "version": "0.4.0"
+  "version": "0.4.1"
 }
