@@ -1,26 +1,26 @@
 {
-  "_from": "fast-glob@^3.1.1",
-  "_id": "fast-glob@3.2.7",
+  "_from": "fast-glob@^3.2.9",
+  "_id": "fast-glob@3.2.10",
   "_inBundle": false,
-  "_integrity": "sha512-rYGMRwip6lUMvYD3BTScMwT1HtAs2d71SMv66Vrxs0IekGZEjhM0pcMfjQPnknBt2zeCwQMEupiN02ZP4DiT1Q==",
+  "_integrity": "sha512-s9nFhFnvR63wls6/kM88kQqDhMu0AfdjqouE2l5GVQPbqLgyFjjU5ry/r2yKsJxpb9Py1EYNqieFrmMaX4v++A==",
   "_location": "/fast-glob",
   "_phantomChildren": {},
   "_requested": {
     "type": "range",
     "registry": true,
-    "raw": "fast-glob@^3.1.1",
+    "raw": "fast-glob@^3.2.9",
     "name": "fast-glob",
     "escapedName": "fast-glob",
-    "rawSpec": "^3.1.1",
+    "rawSpec": "^3.2.9",
     "saveSpec": null,
-    "fetchSpec": "^3.1.1"
+    "fetchSpec": "^3.2.9"
   },
   "_requiredBy": [
     "/globby"
   ],
-  "_resolved": "https://registry.npmjs.org/fast-glob/-/fast-glob-3.2.7.tgz",
-  "_shasum": "fd6cb7a2d7e9aa7a7846111e85a196d6b2f766a1",
-  "_spec": "fast-glob@^3.1.1",
+  "_resolved": "https://registry.npmjs.org/fast-glob/-/fast-glob-3.2.10.tgz",
+  "_shasum": "2734f83baa7f43b7fd41e13bc34438f4ffe284ee",
+  "_spec": "fast-glob@^3.2.9",
   "_where": "C:\\Users\\ronid\\Documents\\CS Final Project\\frontend\\node_modules\\globby",
   "author": {
     "name": "Denis Malinochkin",
@@ -71,7 +71,7 @@
     "typescript": "^3.6.3"
   },
   "engines": {
-    "node": ">=8"
+    "node": ">=8.6.0"
   },
   "files": [
     "out",
@@ -122,5 +122,5 @@
     "watch": "npm run clean && npm run compile -- --sourceMap --watch"
   },
   "typings": "out/index.d.ts",
-  "version": "3.2.7"
+  "version": "3.2.10"
 }
