@@ -1,28 +1,28 @@
 {
-  "_from": "http-errors@1.7.2",
-  "_id": "http-errors@1.7.2",
+  "_from": "http-errors@1.8.1",
+  "_id": "http-errors@1.8.1",
   "_inBundle": false,
-  "_integrity": "sha512-uUQBt3H/cSIVfch6i1EuPNy/YsRSOUBXTVfZ+yR7Zjez3qjBz6i9+i4zjNaoqcoFVI4lQJ5plg63TvGfRSDCRg==",
+  "_integrity": "sha512-Kpk9Sm7NmI+RHhnj6OIWDI1d6fIoFAtFt9RLaTMRlg/8w49juAStsrBgp0Dp4OdxdVbRIeKhtCUvoi/RuAhO4g==",
   "_location": "/http-errors",
   "_phantomChildren": {},
   "_requested": {
     "type": "version",
     "registry": true,
-    "raw": "http-errors@1.7.2",
+    "raw": "http-errors@1.8.1",
     "name": "http-errors",
     "escapedName": "http-errors",
-    "rawSpec": "1.7.2",
+    "rawSpec": "1.8.1",
     "saveSpec": null,
-    "fetchSpec": "1.7.2"
+    "fetchSpec": "1.8.1"
   },
   "_requiredBy": [
     "/body-parser",
     "/raw-body",
     "/send"
   ],
-  "_resolved": "https://registry.npmjs.org/http-errors/-/http-errors-1.7.2.tgz",
-  "_shasum": "4f5029cf13239f31036e5b2e55292bcfbcc85c8f",
-  "_spec": "http-errors@1.7.2",
+  "_resolved": "https://registry.npmjs.org/http-errors/-/http-errors-1.8.1.tgz",
+  "_shasum": "7c3f28577cbc8a207388455dbd62295ed07bd68c",
+  "_spec": "http-errors@1.8.1",
   "_where": "C:\\Users\\ronid\\Documents\\CS Final Project\\frontend\\node_modules\\body-parser",
   "author": {
     "name": "Jonathan Ong",
@@ -45,23 +45,23 @@
   ],
   "dependencies": {
     "depd": "~1.1.2",
-    "inherits": "2.0.3",
-    "setprototypeof": "1.1.1",
+    "inherits": "2.0.4",
+    "setprototypeof": "1.2.0",
     "statuses": ">= 1.5.0 < 2",
-    "toidentifier": "1.0.0"
+    "toidentifier": "1.0.1"
   },
   "deprecated": false,
   "description": "Create HTTP error objects",
   "devDependencies": {
-    "eslint": "5.13.0",
-    "eslint-config-standard": "12.0.0",
-    "eslint-plugin-import": "2.16.0",
-    "eslint-plugin-markdown": "1.0.0",
-    "eslint-plugin-node": "7.0.1",
-    "eslint-plugin-promise": "4.0.1",
-    "eslint-plugin-standard": "4.0.0",
-    "istanbul": "0.4.5",
-    "mocha": "5.2.0"
+    "eslint": "7.32.0",
+    "eslint-config-standard": "14.1.1",
+    "eslint-plugin-import": "2.25.3",
+    "eslint-plugin-markdown": "2.2.1",
+    "eslint-plugin-node": "11.1.0",
+    "eslint-plugin-promise": "5.1.1",
+    "eslint-plugin-standard": "4.1.0",
+    "mocha": "9.1.3",
+    "nyc": "15.1.0"
   },
   "engines": {
     "node": ">= 0.6"
@@ -84,10 +84,10 @@
     "url": "git+https://github.com/jshttp/http-errors.git"
   },
   "scripts": {
-    "lint": "eslint --plugin markdown --ext js,md . && node ./scripts/lint-readme-list.js",
+    "lint": "eslint . && node ./scripts/lint-readme-list.js",
     "test": "mocha --reporter spec --bail",
-    "test-cov": "istanbul cover node_modules/mocha/bin/_mocha -- --reporter dot",
-    "test-travis": "istanbul cover node_modules/mocha/bin/_mocha --report lcovonly -- --reporter dot"
+    "test-ci": "nyc --reporter=lcov --reporter=text npm test",
+    "test-cov": "nyc --reporter=html --reporter=text npm test"
   },
-  "version": "1.7.2"
+  "version": "1.8.1"
 }
