@@ -1,26 +1,26 @@
 {
-  "_from": "node-forge@^0.10.0",
-  "_id": "node-forge@0.10.0",
+  "_from": "node-forge@^1.2.0",
+  "_id": "node-forge@1.2.1",
   "_inBundle": false,
-  "_integrity": "sha512-PPmu8eEeG9saEUvI97fm4OYxXVB6bFvyNTyiUOBichBpFG8A1Ljw3bY62+5oOjDEMHRnd0Y7HQ+x7uzxOzC6JA==",
+  "_integrity": "sha512-Fcvtbb+zBcZXbTTVwqGA5W+MKBj56UjVRevvchv5XrcyXbmNdesfZL37nlcWOfpgHhgmxApw3tQbTr4CqNmX4w==",
   "_location": "/node-forge",
   "_phantomChildren": {},
   "_requested": {
     "type": "range",
     "registry": true,
-    "raw": "node-forge@^0.10.0",
+    "raw": "node-forge@^1.2.0",
     "name": "node-forge",
     "escapedName": "node-forge",
-    "rawSpec": "^0.10.0",
+    "rawSpec": "^1.2.0",
     "saveSpec": null,
-    "fetchSpec": "^0.10.0"
+    "fetchSpec": "^1.2.0"
   },
   "_requiredBy": [
     "/selfsigned"
   ],
-  "_resolved": "https://registry.npmjs.org/node-forge/-/node-forge-0.10.0.tgz",
-  "_shasum": "32dea2afb3e9926f02ee5ce8794902691a676bf3",
-  "_spec": "node-forge@^0.10.0",
+  "_resolved": "https://registry.npmjs.org/node-forge/-/node-forge-1.2.1.tgz",
+  "_shasum": "82794919071ef2eb5c509293325cec8afd0fd53c",
+  "_spec": "node-forge@^1.2.0",
   "_where": "C:\\Users\\ronid\\Documents\\CS Final Project\\frontend\\node_modules\\selfsigned",
   "author": {
     "name": "Digital Bazaar, Inc.",
@@ -61,8 +61,8 @@
     "browserify": "^16.5.2",
     "commander": "^2.20.0",
     "cross-env": "^5.2.1",
-    "eslint": "^7.8.1",
-    "eslint-config-digitalbazaar": "^2.5.0",
+    "eslint": "^7.27.0",
+    "eslint-config-digitalbazaar": "^2.8.0",
     "express": "^4.16.2",
     "karma": "^4.4.1",
     "karma-browserify": "^7.0.0",
@@ -87,7 +87,7 @@
     "worker-loader": "^2.0.0"
   },
   "engines": {
-    "node": ">= 6.0.0"
+    "node": ">= 6.13.0"
   },
   "files": [
     "lib/*.js",
@@ -143,16 +143,18 @@
   "scripts": {
     "build": "webpack",
     "coverage": "rm -rf coverage && nyc --reporter=lcov --reporter=text-summary npm test",
+    "coverage-ci": "rm -rf coverage && nyc --reporter=lcovonly npm test",
     "coverage-report": "nyc report",
     "lint": "eslint *.js lib/*.js tests/*.js tests/**/*.js examples/*.js flash/*.js",
     "prepublish": "npm run build",
-    "test": "cross-env NODE_ENV=test mocha -t 30000 -R ${REPORTER:-spec} tests/unit/index.js",
+    "test": "npm run test-node",
     "test-build": "webpack --config webpack-tests.config.js",
     "test-karma": "karma start",
     "test-karma-sauce": "karma start karma-sauce.conf",
+    "test-node": "cross-env NODE_ENV=test mocha -t 30000 -R ${REPORTER:-spec} tests/unit/index.js",
     "test-server": "node tests/server.js",
     "test-server-webid": "node tests/websockets/server-webid.js",
     "test-server-ws": "node tests/websockets/server-ws.js"
   },
-  "version": "0.10.0"
+  "version": "1.2.1"
 }
