@@ -1,27 +1,27 @@
 {
-  "_from": "qs@6.7.0",
-  "_id": "qs@6.7.0",
+  "_from": "qs@6.9.6",
+  "_id": "qs@6.9.6",
   "_inBundle": false,
-  "_integrity": "sha512-VCdBRNFTX1fyE7Nb6FYoURo/SPe62QCaAyzJvUjwRaIsc+NePBEniHlvxFmmX56+HZphIGtV0XeCirBtpDrTyQ==",
+  "_integrity": "sha512-TIRk4aqYLNoJUbd+g2lEdz5kLWIuTMRagAXxl78Q0RiVjAOugHmeKNGdd3cwo/ktpf9aL9epCfFqWDEKysUlLQ==",
   "_location": "/qs",
   "_phantomChildren": {},
   "_requested": {
     "type": "version",
     "registry": true,
-    "raw": "qs@6.7.0",
+    "raw": "qs@6.9.6",
     "name": "qs",
     "escapedName": "qs",
-    "rawSpec": "6.7.0",
+    "rawSpec": "6.9.6",
     "saveSpec": null,
-    "fetchSpec": "6.7.0"
+    "fetchSpec": "6.9.6"
   },
   "_requiredBy": [
     "/body-parser",
     "/express"
   ],
-  "_resolved": "https://registry.npmjs.org/qs/-/qs-6.7.0.tgz",
-  "_shasum": "41dc1a015e3d581f1621776be31afb2876a9b1bc",
-  "_spec": "qs@6.7.0",
+  "_resolved": "https://registry.npmjs.org/qs/-/qs-6.9.6.tgz",
+  "_shasum": "26ed3c8243a431b2924aca84cc90471f35d5a0ee",
+  "_spec": "qs@6.9.6",
   "_where": "C:\\Users\\ronid\\Documents\\CS Final Project\\frontend\\node_modules\\express",
   "bugs": {
     "url": "https://github.com/ljharb/qs/issues"
@@ -38,23 +38,34 @@
   "deprecated": false,
   "description": "A querystring parser that supports nesting and arrays, with a depth limit",
   "devDependencies": {
-    "@ljharb/eslint-config": "^13.1.1",
-    "browserify": "^16.2.3",
-    "covert": "^1.1.1",
-    "editorconfig-tools": "^0.1.1",
-    "eslint": "^5.15.3",
-    "evalmd": "^0.0.17",
+    "@ljharb/eslint-config": "^17.3.0",
+    "aud": "^1.1.3",
+    "browserify": "^16.5.2",
+    "eclint": "^2.8.1",
+    "eslint": "^7.17.0",
+    "evalmd": "^0.0.19",
     "for-each": "^0.3.3",
-    "iconv-lite": "^0.4.24",
-    "mkdirp": "^0.5.1",
-    "object-inspect": "^1.6.0",
+    "has-symbols": "^1.0.1",
+    "iconv-lite": "^0.5.1",
+    "mkdirp": "^0.5.5",
+    "nyc": "^10.3.2",
+    "object-inspect": "^1.9.0",
     "qs-iconv": "^1.0.4",
-    "safe-publish-latest": "^1.1.2",
+    "safe-publish-latest": "^1.1.4",
     "safer-buffer": "^2.1.2",
-    "tape": "^4.10.1"
+    "tape": "^5.1.1"
   },
   "engines": {
     "node": ">=0.6"
+  },
+  "funding": {
+    "url": "https://github.com/sponsors/ljharb"
+  },
+  "greenkeeper": {
+    "ignore": [
+      "iconv-lite",
+      "mkdirp"
+    ]
   },
   "homepage": "https://github.com/ljharb/qs",
   "keywords": [
@@ -73,15 +84,15 @@
     "url": "git+https://github.com/ljharb/qs.git"
   },
   "scripts": {
-    "coverage": "covert test",
     "dist": "mkdirp dist && browserify --standalone Qs lib/index.js > dist/qs.js",
     "lint": "eslint lib/*.js test/*.js",
-    "postlint": "editorconfig-tools check * lib/* test/*",
+    "postlint": "eclint check * lib/* test/*",
+    "posttest": "aud --production",
     "prepublish": "safe-publish-latest && npm run dist",
     "pretest": "npm run --silent readme && npm run --silent lint",
     "readme": "evalmd README.md",
-    "test": "npm run --silent coverage",
-    "tests-only": "node test"
+    "test": "npm run tests-only",
+    "tests-only": "nyc tape 'test/**/*.js'"
   },
-  "version": "6.7.0"
+  "version": "6.9.6"
 }
