@@ -1,19 +1,34 @@
-import React, {Component} from "react";
+import React, {useState} from "react";
 import Navigation from '../components/Navigation';
 import CertificationTable from "../components/team/CertificationTable";
+import Search from "../components/team/Search";
+import { UserCertificationsActions } from '../services/api/userCertificationsActions'
 
-export default class Team extends Component{
-  constructor(props) {
-    super(props);
+function Team(){
+
+  let [data, setData] = useState([]);
+  let [firstRender, setFirstRender] = useState(true);
+
+  if (firstRender){
+    setFirstRender(false);
+    getData();
   }
 
-  render() {
-    return (
-      <div>
-        <Navigation />
-        <h1>Team Page</h1>
-        <CertificationTable />
-      </div>
-    );
+  async function getData(){ 
+    let certifications = new UserCertificationsActions();
+    await certifications.printableData().then(response => {
+      setData(response);
+    })
   }
+
+  return (
+    <div>
+      <Navigation />
+      <h1>Team Page</h1>
+      <Search data={data} setData={setData} />
+      <CertificationTable data={data} />
+    </div>
+  );
 }
+
+export default Team;
